requirements.txt
pandas
psycopg2
python-dotenv==0.11.0
SQLAlchemy
